+++ services/yellow/src/api/Components/OrderPaymentManger.cs
@@ -1,0 +1,13 @@
+using System;
+using System.Threading.Tasks;
+using artiso.AdsdHotel.Yellow.Api.Validation;
+using artiso.AdsdHotel.Yellow.Contracts.Commands;
+using artiso.AdsdHotel.Yellow.Contracts.Models;
+
+namespace artiso.AdsdHotel.Yellow.Api.Components
+{
+    public class OrderPaymentManger
+    {
+        
+    }
+}
