+++ services/yellow/src/api/Handlers/ChargeForOrderFullAmountHandler.cs
@@ -1,16 +1,14 @@
-using System;
-using System.Linq;
 using System.Threading.Tasks;
+using artiso.AdsdHotel.Yellow.Api.Handlers.Templates;
 using artiso.AdsdHotel.Yellow.Api.Services;
 using artiso.AdsdHotel.Yellow.Api.Validation;
-using artiso.AdsdHotel.Yellow.Contracts;
 using artiso.AdsdHotel.Yellow.Contracts.Commands;
 using artiso.AdsdHotel.Yellow.Contracts.Models;
 using artiso.AdsdHotel.Yellow.Events;
 
 namespace artiso.AdsdHotel.Yellow.Api.Handlers
 {
-    public class ChargeForOrderFullAmountHandler : AbstractHandler<ChargeForOrderFullAmountRequest, OrderFullAmountCharged>
+    public class ChargeForOrderFullAmountHandler : AbstractChargeHandler<ChargeForOrderFullAmountRequest, OrderFullAmountCharged>
     {
         private readonly IOrderService _orderService;
         private readonly ICreditCardPaymentService _creditCardPaymentService;
@@ -25,32 +23,12 @@ namespace artiso.AdsdHotel.Yellow.Api.Handlers
         {
             var order = await Ensure(message, r => _orderService.FindOneById(r.OrderId));
 
-            var activePaymentMethod = order.OrderPaymentMethods.LastOrDefault();
+            var amountToPay = order.Price.Amount;
 
             if (message.AlternativePaymentMethod is not null)
-            {
-                var creditCard = message.AlternativePaymentMethod.CreditCard;
-                var chargeResult = await _creditCardPaymentService.Charge(order.Price.Amount, creditCard!);
-
-                if (!chargeResult.IsSuccess)
-                    throw new InvalidOperationException(chargeResult.Exception?.Message);
-
-                if (chargeResult.AuthorizePaymentToken is null)
-                    throw new InvalidOperationException("Charge result does not contain token");
-
-                var orderPaymentMethod = new OrderPaymentMethod(creditCard.GetOrderCreditCard(chargeResult.AuthorizePaymentToken));
-                
-                await _orderService.AddPaymentMethod(order, orderPaymentMethod);
-                
-                return new OrderFullAmountCharged(message.OrderId);
-            }
-            
-            var creditCardProviderPaymentToken = activePaymentMethod?.CreditCard.ProviderPaymentToken;
-            
-            var r = await _creditCardPaymentService.Charge(order.Price.Amount, creditCardProviderPaymentToken!);
-            
-            if (!r.IsSuccess)
-                throw new InvalidOperationException(r.Exception?.Message);
+                await ChargeOrder(order, amountToPay, message.AlternativePaymentMethod);
+            else
+                await ChargeOrder(order, amountToPay);
             
             return new OrderFullAmountCharged(message.OrderId);
         }
@@ -62,10 +40,26 @@ namespace artiso.AdsdHotel.Yellow.Api.Handlers
                 .HasData(m => m.OrderId, $"Missing {nameof(ChargeForOrderFullAmountRequest.OrderId)}");
             
             if (message.AlternativePaymentMethod is not null)
-                return v.That(m => m.AlternativePaymentMethod.Validate().PaymentMethodIsValid(p => p).IsValid(),
+                return v.That(m => m.AlternativePaymentMethod.Validate()
+                        .PaymentMethodIsValid(p => p!).IsValid(),
                     $"{nameof(ChargeForOrderFullAmountRequest.AlternativePaymentMethod)} must have a valid creditcard");
 
             return v;
+        }
+
+        protected override Task AddPaymentMethod(Order order, OrderPaymentMethod paymentMethod)
+        {
+            return _orderService.AddPaymentMethod(order, paymentMethod);
+        }
+
+        protected override Task<ChargeResult> Charge(decimal amount, string paymentToken)
+        {
+            return _creditCardPaymentService.Charge(amount, paymentToken);
+        }
+
+        protected override Task<ChargeResult> Charge(decimal amount, CreditCard creditCard)
+        {
+            return _creditCardPaymentService.Charge(amount, creditCard);
         }
     }
 }
