+++ services/yellow/src/api/artiso.AdsdHotel.Yellow.Api.csproj
